+++ packages/core/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  collectCoverage: true,
+  clearMocks: true,
+  fakeTimers: { enableGlobally: true },
+  coverageDirectory: 'coverage',
+  silent: true,
+  verbose: false,
+}